requirements.txt
python-telegram-bot==20.6
apscheduler==3.10.4
ccxt==4.2.97
pandas==2.3.1
pandas-ta==0.3.14b0
matplotlib==3.9.2
python-dotenv==1.0.1
pytz==2025.2